requirements.txt
requests
selenium
beautifulsoup4
hazelcast-python-client
kafka-python
feed-utils==0.3.9
flask-classy
flask
